extracting paths operations in main.py to specific files under routers folder
